[BACKLOG-18829] Reorganize AvroOutputMeta in preparation of native spark operation
